# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "EdgeCraft Swagger API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/clouds/{cloudId}/clusters": {
            "get": {
                "description": "전체 클러스터 리스트 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster"
                ],
                "summary": "GetClusterList",
                "operationId": "GetClusterList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "클러스터 추가 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster"
                ],
                "summary": "SetCluster",
                "operationId": "SetCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Openstack Cluster Info",
                        "name": "OpenstackClusterSet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OpenstackClusterSet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
        "/clouds/{cloudId}/clusters/{clusterId}": {
            "get": {
                "description": "클러스터 상세 조회 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster"
                ],
                "summary": "GetCluster",
                "operationId": "GetCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "put": {
                "description": "클러스터 수정 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster"
                ],
                "summary": "UpdateCluster",
                "operationId": "UpdateCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Openstack Cluster Info",
                        "name": "OpenstackClusterSet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OpenstackClusterSet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "저장된 클러스터 정보를 이용해서 Provision 처리 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster"
                ],
                "summary": "ProvisioningCluster",
                "operationId": "ProvisioningCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "delete": {
                "description": "클러스터 삭제 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster"
                ],
                "summary": "DeleteCluster",
                "operationId": "DeleteCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
...
        "/clouds/{cloudId}/clusters/{clusterId}/upgrade": {
            "post": {
                "description": "저장된 클러스터 정보를 이용해서 K8s Version Upgrade 처리 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster"
                ],
                "summary": "UpgradeClusterK8sVersion",
                "operationId": "UpgradeClusterK8sVersion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Openstack Cluster K8s Upgrade Info",
                        "name": "K8sUpgradeInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.K8sUpgradeInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.EtcdInfo": {
            "type": "object",
            "properties": {
                "ca_file": {
                    "type": "string",
                    "example": ""
                },
                "cert_file": {
                    "type": "string",
                    "example": ""
                },
                "endpoints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UrlInfo"
                    }
                },
                "key_file": {
                    "type": "string",
                    "example": ""
                },
                "use_external_etcd": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "model.EtcdStorageInfo": {
            "type": "object",
            "properties": {
                "etcd": {
                    "$ref": "#/definitions/model.EtcdInfo"
                },
                "storage_class": {
                    "$ref": "#/definitions/model.StorageClassInfo"
                }
            }
        },
...
        "model.K8sUpgradeInfo": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "model.KubernetesInfo": {
            "type": "object",
            "properties": {
                "bootstrap_provider": {
                    "type": "integer",
                    "example": 1
                },
                "cp_kubeadm_extra_config": {
                    "$ref": "#/definitions/model.ExtraConfig"
                },
                "pod_cidr": {
                    "type": "string",
                    "example": "10.96.0.1/12"
                },
                "svc_cidr": {
                    "type": "string",
                    "example": "10.96.0.0/12"
                },
                "svc_domain": {
                    "type": "string",
                    "example": "cluster.local"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                },
                "worker_kubeadm_extra_config": {
                    "$ref": "#/definitions/model.ExtraConfig"
                }
            }
        },
...
        "model.OpenstackClusterInfo": {
            "type": "object",
            "properties": {
                "cluster_uid": {
                    "type": "string",
                    "example": ""
                },
                "created": {
                    "type": "string"
                },
                "desc": {
                    "type": "string",
                    "example": "Openstack Test Cluster #1"
                },
                "name": {
                    "type": "string",
                    "example": "os-test-1"
                },
                "namespace": {
                    "type": "string",
                    "example": "default"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "model.OpenstackClusterSet": {
            "type": "object",
            "properties": {
                "cluster": {
                    "$ref": "#/definitions/model.OpenstackClusterInfo"
                },
                "etcd_storage": {
                    "$ref": "#/definitions/model.EtcdStorageInfo"
                },
                "k8s": {
                    "$ref": "#/definitions/model.KubernetesInfo"
                },
                "nodes": {
                    "$ref": "#/definitions/model.OpenstackNodeSetInfo"
                },
                "openstack": {
                    "$ref": "#/definitions/model.OpenstackInfo"
                },
                "save_only": {
                    "type": "boolean"
                }
            }
        },
        "model.OpenstackInfo": {
            "type": "object",
            "properties": {
                "api_server_floating_ip": {
                    "type": "string",
                    "example": ""
                },
                "bastion_flavor": {
                    "type": "string",
                    "example": ""
                },
                "bastion_floating_ip": {
                    "type": "string",
                    "example": ""
                },
                "bastion_image_name": {
                    "type": "string",
                    "example": ""
                },
                "bastion_ssh_key_name": {
                    "type": "string",
                    "example": ""
                },
                "dns_nameservers": {
                    "type": "string",
                    "example": "168.126.63.1"
                },
                "external_network_id": {
                    "description": "public",
                    "type": "string",
                    "example": ""
                },
                "failure_domain": {
                    "description": "nova",
                    "type": "string",
                    "example": ""
                },
                "image_name": {
                    "type": "string",
                    "example": "ubuntu-2004-kube-v1.23.3"
                },
                "node_cidr": {
                    "type": "string",
                    "example": "10.96.0.0/24"
                },
                "openstack_cloud": {
                    "type": "string",
                    "example": "openstack"
                },
                "openstack_cloud_cacert_b64": {
                    "type": "string",
                    "example": "Cg=="
                },
                "openstack_cloud_provider_conf_b64": {
                    "type": "string",
                    "example": "W0dsb2JhbF0KYXV0aC11cmw9aHR0cDovLzE5Mi4xNjguNzcuMTEvaWRlbnRpdHkKdXNlcm5hbWU9InN1bm1pIgpwYXNzd29yZD0iZmtmZms0NDgiCnRlbmFudC1pZD0iNTQyZTdhMDRmNjkxNDgyOWI0M2U3N2Y5ZWYxMmI3NzkiCnRlbmFudC1uYW1lPSJlZGdlY3JhZnQiCmRvbWFpbi1uYW1lPSJEZWZhdWx0IgpyZWdpb249IlJlZ2lvbk9uZSIK"
                },
                "openstack_cloud_yaml_b64": {
                    "type": "string",
                    "example": "Y2xvdWRzOgogIG9wZW5zdGFjazoKICAgIGF1dGg6CiAgICAgIGF1dGhfdXJsOiBodHRwOi8vMTkyLjE2OC43Ny4xMS9pZGVudGl0eQogICAgICB1c2VybmFtZTogInN1bm1pIgogICAgICBwYXNzd29yZDogImZrZmZrNDQ4IgogICAgICBwcm9qZWN0X2lkOiA1NDJlN2EwNGY2OTE0ODI5YjQzZTc3ZjllZjEyYjc3OQogICAgICBwcm9qZWN0X25hbWU6ICJlZGdlY3JhZnQiCiAgICAgIHVzZXJfZG9tYWluX25hbWU6ICJEZWZhdWx0IgogICAgcmVnaW9uX25hbWU6ICJSZWdpb25PbmUiCiAgICBpbnRlcmZhY2U6ICJwdWJsaWMiCiAgICBpZGVudGl0eV9hcGlfdmVyc2lvbjogMwo="
                },
                "ssh_key_name": {
                    "type": "string",
                    "example": "sunmi"
                },
                "use_bastion_host": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "model.OpenstackNodeSetInfo": {
            "type": "object",
            "properties": {
                "master_sets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NodeSetInfo"
                    }
                },
                "use_loadbalancer": {
                    "type": "boolean",
                    "example": false
                },
                "worker_sets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NodeSetInfo"
                    }
                }
            }
        },
...
        "model.UrlInfo": {
            "type": "object",
            "properties": {
                "ip_address": {
                    "type": "string",
                    "example": ""
                },
                "port": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "response.ReturnData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "isError": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}